Identify MPO as Municipal Police Officers
Municipal police officers are explained in this article from the

https://laapoa.com/2021/08/laapoa-member-spotlight-officer-manuel-elorriaga/

"Elorriaga became a municipal police officer (MPO) when the General
Service Police Department (LAGSPD) was dissolved by the L.A. City
Council and absorbed into the LAPD in 2012.

"For those unsure where MPOs fit within the city’s policing services,
Elorriaga offers some clarification. 'An MPO is a former LAGSPD officer
who did not complete a 1014 lateral transfer into the LAPD,” he
explains. “As a result, those of us who didn’t complete the 1014
transfer were kept in our PC classification, but we complete all
training and are kept to the same standards as any LAPD officer.'”